% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_mSig.R
\name{compute_mSig}
\alias{compute_mSig}
\title{Compute the predicted EGFR mSig score using predefined LUAD centroids}
\usage{
compute_mSig(
  centroids = NULL,
  expr,
  meanCol = "all",
  fname = "Pred_mSig",
  export = TRUE
)
}
\arguments{
\item{centroids}{All_subtypes_centroids (1020 genes x 10 means), z-score space}

\item{expr}{gene expression matrix (gene symbols x sample IDs), values = RSEM TPM}
}
\value{
pred.ind (1=WT-like, 2=mt-like), pred.score (distance)
}
\description{
Compute the predicted EGFR mSig score using predefined LUAD centroids
}
