% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mSig.R
\name{plot_mSig}
\alias{plot_mSig}
\title{Plot the distribution of EGFR mSig distance by class}
\usage{
plot_mSig(res)
}
\arguments{
\item{res}{output object returned by \code{compute_mSig()}}
}
\value{
violin and boxplot
}
\description{
Plot the distribution of EGFR mSig distance by class
}
